Move admin-subnet commands to support (#3949)
