+++ cmd/admin-subnet.go
@@ -19,6 +19,7 @@ package cmd
 
 import (
 	"github.com/minio/cli"
+	"github.com/minio/pkg/console"
 )
 
 var subnetHealthSubcommands = []cli.Command{
@@ -27,18 +28,18 @@ var subnetHealthSubcommands = []cli.Command{
 }
 
 var adminSubnetCmd = cli.Command{
-	Name:            "subnet",
-	Usage:           "Subnet related commands",
-	Action:          mainAdminSubnet,
-	Before:          setGlobalsFromContext,
-	Flags:           globalFlags,
-	Subcommands:     subnetHealthSubcommands,
-	HideHelpCommand: true,
+	Name:        "subnet",
+	Usage:       "Subnet related commands",
+	Action:      mainAdminSubnet,
+	Before:      setGlobalsFromContext,
+	Flags:       globalFlags,
+	Subcommands: subnetHealthSubcommands,
+	Hidden:      true,
 }
 
 // mainAdminSubnet is the handle for "mc admin subnet" command.
 func mainAdminSubnet(ctx *cli.Context) error {
-	commandNotFound(ctx, subnetHealthSubcommands)
+	console.Infoln("Please use 'mc support'")
 	return nil
 	// Sub-commands like "health", "register" have their own main.
 }
